requirements.txt
asgiref==3.2.10
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
coverage==5.2.1
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-filter==2.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
factory-boy==3.0.1
Faker==4.1.2
idna==2.10
inflection==0.5.1
iniconfig==1.0.1
mock==4.0.2
more-itertools==8.4.0
oauthlib==3.1.0
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
pytest-cov==2.10.0
pytest-django==3.9.0
pytest-factoryboy==2.0.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
urllib3==1.25.10
